Refactored code to group by algorithm
