+++ src/jose/jwk.ts
@@ -1,39 +1,12 @@
-import * as ed25519 from '@noble/ed25519';
-import base64url from 'base64url';
-import type { JwkEd25519Private, JwkEd25519Public } from './jwk-types';
+import type { JwkEd25519Private, JwkEd25519Public } from './ed25519';
+import type { JwkSecp256k1Private, JwkSecp256k1Public } from './secp256k1';
 
 /**
- * Class containing JWK related operations.
+ * A supported private key in JWK format.
  */
-export class Jwk {
+export type JwkPrivate = JwkSecp256k1Private | JwkEd25519Private;
 
-  /**
-   * Generates ED25519 public-private key pair.
-   * @returns Public and private ED25519 keys in JWK format.
-   */
-  public static async generateEd25519KeyPair (): Promise<{
-    publicKeyJwk: JwkEd25519Public,
-    privateKeyJwk: JwkEd25519Private
-  }> {
-    const privateKeyUint8Array = ed25519.utils.randomPrivateKey();
-    const publicKeyUint8Array = await ed25519.getPublicKey(privateKeyUint8Array);
-    const privateKeyBuffer = Buffer.from(privateKeyUint8Array);
-    const publicKeyBuffer = Buffer.from(publicKeyUint8Array);
-
-    const d = base64url.encode(privateKeyBuffer);
-    const x = base64url.encode(publicKeyBuffer);
-
-    const publicKeyJwk: JwkEd25519Public = {
-      kty : 'OKP',
-      crv : 'Ed25519',
-      x
-    };
-
-    const privateKeyJwk: Required<JwkEd25519Private> = {
-      ...publicKeyJwk,
-      d
-    };
-
-    return { publicKeyJwk, privateKeyJwk};
-  }
-}
+/**
+ * A supported public key in JWK format.
+ */
+export type JwkPublic = JwkSecp256k1Public | JwkEd25519Public;
